Borrowing Events Edited
Most of the borrowing Events already created, but lots of title ID’s
missing and dates not filled in as written so I hopefully fixed that.

One thing to note—I don’t know if it’s a problem but the following
titles whose ID’s were missing were NOT listed as ever being borrowed
by Felder on the ‘borrowed-titles’ xml file.